GPTQ analysis and plotting
Through the code in this commit, we found that activation order adjusted gptq actually sort the hessian in decending order, hoping for the larger errors to be easily adjusted by the large amount of untrained weights. THis is a sound strategy for constant constant precision and scaling quantization. However, this could be improved by mix-precision quant on the columns, and dynamic scaling adjustment based on the changes to latter layers